Updated demo app to automatically adapt to the underlying platform (Resolves #436)
