+++ super_editor/example/lib/demos/example_editor/example_editor.dart
@@ -1,3 +1,4 @@
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:super_editor/super_editor.dart';
 
@@ -17,12 +18,13 @@ class _ExampleEditorState extends State<ExampleEditor> {
   final GlobalKey _docLayoutKey = GlobalKey();
 
   late Document _doc;
-  DocumentEditor? _docEditor;
-  DocumentComposer? _composer;
+  late DocumentEditor _docEditor;
+  late DocumentComposer _composer;
+  late CommonEditorOperations _docOps;
 
-  FocusNode? _editorFocusNode;
+  late FocusNode _editorFocusNode;
 
-  ScrollController? _scrollController;
+  late ScrollController _scrollController;
 
   OverlayEntry? _formatBarOverlayEntry;
   final _selectionAnchor = ValueNotifier<Offset?>(null);
@@ -33,6 +35,11 @@ class _ExampleEditorState extends State<ExampleEditor> {
     _doc = createInitialDocument()..addListener(_hideOrShowToolbar);
     _docEditor = DocumentEditor(document: _doc as MutableDocument);
     _composer = DocumentComposer()..addListener(_hideOrShowToolbar);
+    _docOps = CommonEditorOperations(
+      editor: _docEditor,
+      composer: _composer,
+      documentLayoutResolver: () => _docLayoutKey.currentState as DocumentLayout,
+    );
     _editorFocusNode = FocusNode();
     _scrollController = ScrollController()..addListener(_hideOrShowToolbar);
   }
@@ -43,14 +50,20 @@ class _ExampleEditorState extends State<ExampleEditor> {
       _formatBarOverlayEntry!.remove();
     }
 
-    _scrollController!.dispose();
-    _editorFocusNode!.dispose();
-    _composer!.dispose();
+    _scrollController.dispose();
+    _editorFocusNode.dispose();
+    _composer.dispose();
     super.dispose();
   }
 
   void _hideOrShowToolbar() {
-    final selection = _composer!.selection;
+    if (_gestureMode != DocumentGestureMode.mouse) {
+      // We only add our own toolbar when using mouse. On mobile, a bar
+      // is rendered for us.
+      return;
+    }
+
+    final selection = _composer.selection;
     if (selection == null) {
       // Nothing is selected. We don't want to show a toolbar
       // in this case.
@@ -114,7 +127,7 @@ class _ExampleEditorState extends State<ExampleEditor> {
       }
 
       final docBoundingBox = (_docLayoutKey.currentState as DocumentLayout)
-          .getRectForSelection(_composer!.selection!.base, _composer!.selection!.extent)!;
+          .getRectForSelection(_composer.selection!.base, _composer.selection!.extent)!;
       final docBox = _docLayoutKey.currentContext!.findRenderObject() as RenderBox;
       final overlayBoundingBox = Rect.fromPoints(
         docBox.localToGlobal(docBoundingBox.topLeft, ancestor: context.findRenderObject()),
@@ -145,19 +158,98 @@ class _ExampleEditorState extends State<ExampleEditor> {
     // I tried explicitly unfocus()'ing the URL textfield
     // in the toolbar but it didn't return focus to the
     // editor. I'm not sure why.
-    _editorFocusNode!.requestFocus();
+    _editorFocusNode.requestFocus();
   }
+
+  DocumentGestureMode get _gestureMode {
+    switch (defaultTargetPlatform) {
+      case TargetPlatform.android:
+        return DocumentGestureMode.android;
+      case TargetPlatform.iOS:
+        return DocumentGestureMode.iOS;
+      case TargetPlatform.fuchsia:
+      case TargetPlatform.linux:
+      case TargetPlatform.macOS:
+      case TargetPlatform.windows:
+        return DocumentGestureMode.mouse;
+    }
+  }
+
+  bool get _isMobile => _gestureMode != DocumentGestureMode.mouse;
+
+  DocumentInputSource get _inputSource {
+    switch (defaultTargetPlatform) {
+      case TargetPlatform.android:
+      case TargetPlatform.iOS:
+        return DocumentInputSource.ime;
+      case TargetPlatform.fuchsia:
+      case TargetPlatform.linux:
+      case TargetPlatform.macOS:
+      case TargetPlatform.windows:
+        return DocumentInputSource.keyboard;
+    }
+  }
+
+  void _cut() => _docOps.cut();
+  void _copy() => _docOps.copy();
+  void _paste() => _docOps.paste();
+  void _selectAll() => _docOps.selectAll();
 
   @override
   Widget build(BuildContext context) {
+    return Column(
+      children: [
+        Expanded(
+          child: _buildEditor(),
+        ),
+        if (_isMobile)
+          MultiListenableBuilder(
+            listenables: <Listenable>{
+              _doc,
+              _composer.selectionNotifier,
+            },
+            builder: (_) => _buildMountedToolbar(),
+          ),
+      ],
+    );
+  }
+
+  Widget _buildEditor() {
     return SuperEditor(
-      editor: _docEditor!,
+      editor: _docEditor,
       composer: _composer,
       focusNode: _editorFocusNode,
       scrollController: _scrollController,
       documentLayoutKey: _docLayoutKey,
       maxWidth: 600, // arbitrary choice for maximum width
       padding: const EdgeInsets.symmetric(vertical: 56, horizontal: 24),
+      gestureMode: _gestureMode,
+      inputSource: _inputSource,
+      androidToolbarBuilder: (_) => AndroidTextEditingFloatingToolbar(
+        onCutPressed: _cut,
+        onCopyPressed: _copy,
+        onPastePressed: _paste,
+        onSelectAllPressed: _selectAll,
+      ),
+      iOSToolbarBuilder: (_) => IOSTextEditingFloatingToolbar(
+        onCutPressed: _cut,
+        onCopyPressed: _copy,
+        onPastePressed: _paste,
+      ),
+    );
+  }
+
+  Widget _buildMountedToolbar() {
+    final selection = _composer.selection;
+
+    if (selection == null) {
+      return const SizedBox();
+    }
+
+    return KeyboardEditingToolbar(
+      document: _doc,
+      composer: _composer,
+      commonOps: _docOps,
     );
   }
 }
