+++ super_editor/lib/src/default_editor/document_keyboard_actions.dart
@@ -1,24 +1,15 @@
-import 'dart:math';
-
 import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 import 'package:super_editor/src/core/document.dart';
-import 'package:super_editor/src/core/document_composer.dart';
-import 'package:super_editor/src/core/document_editor.dart';
 import 'package:super_editor/src/core/document_layout.dart';
 import 'package:super_editor/src/core/document_selection.dart';
 import 'package:super_editor/src/core/edit_context.dart';
 import 'package:super_editor/src/default_editor/attributions.dart';
-import 'package:super_editor/src/default_editor/common_editor_operations.dart';
 import 'package:super_editor/src/infrastructure/_logging.dart';
-import 'package:super_editor/src/infrastructure/attributed_text.dart';
 import 'package:super_editor/src/infrastructure/keyboard.dart';
 
-import 'box_component.dart';
 import 'document_input_keyboard.dart';
-import 'multi_node_editing.dart';
 import 'paragraph.dart';
-import 'selection_binary.dart';
 import 'text.dart';
 
 final _log = Logger(scope: 'document_keyboard_actions.dart');
@@ -47,31 +38,7 @@ ExecutionInstruction pasteWhenCmdVIsPressed({
   }
 
   _log.log('pasteWhenCmdVIsPressed', 'Pasting clipboard content...');
-  DocumentPosition pastePosition = editContext.composer.selection!.extent;
-
-  // Delete all currently selected content.
-  if (!editContext.composer.selection!.isCollapsed) {
-    pastePosition = CommonEditorOperations.getDocumentPositionAfterExpandedDeletion(
-      document: editContext.editor.document,
-      selection: editContext.composer.selection!,
-    );
-
-    // Delete the selected content.
-    editContext.editor.executeCommand(
-      DeleteSelectionCommand(documentSelection: editContext.composer.selection!),
-    );
-
-    editContext.composer.selection = DocumentSelection.collapsed(position: pastePosition);
-  }
-
-  // TODO: figure out a general approach for asynchronous behaviors that
-  //       need to be carried out in response to user input.
-  _paste(
-    document: editContext.editor.document,
-    editor: editContext.editor,
-    composer: editContext.composer,
-    pastePosition: pastePosition,
-  );
+  editContext.commonOps.paste();
 
   return ExecutionInstruction.haltExecution;
 }
@@ -86,139 +53,6 @@ ExecutionInstruction selectAllWhenCmdAIsPressed({
 
   final didSelectAll = editContext.commonOps.selectAll();
   return didSelectAll ? ExecutionInstruction.haltExecution : ExecutionInstruction.continueExecution;
-}
-
-Future<void> _paste({
-  required Document document,
-  required DocumentEditor editor,
-  required DocumentComposer composer,
-  required DocumentPosition pastePosition,
-}) async {
-  final content = (await Clipboard.getData('text/plain'))?.text ?? '';
-  _log.log('_paste', 'Content from clipboard: $content');
-
-  editor.executeCommand(
-    _PasteEditorCommand(
-      content: content,
-      pastePosition: pastePosition,
-      composer: composer,
-    ),
-  );
-}
-
-class _PasteEditorCommand implements EditorCommand {
-  _PasteEditorCommand({
-    required String content,
-    required DocumentPosition pastePosition,
-    required DocumentComposer composer,
-  })  : _content = content,
-        _pastePosition = pastePosition,
-        _composer = composer;
-
-  final String _content;
-  final DocumentPosition _pastePosition;
-  final DocumentComposer _composer;
-
-  @override
-  void execute(Document document, DocumentEditorTransaction transaction) {
-    final splitContent = _content.split('\n\n');
-    _log.log('_PasteEditorCommand', 'Split content:');
-    for (final piece in splitContent) {
-      _log.log('_PasteEditorCommand', ' - "$piece"');
-    }
-
-    final currentNodeWithSelection = document.getNodeById(_pastePosition.nodeId);
-
-    DocumentPosition? newSelectionPosition;
-
-    if (currentNodeWithSelection is TextNode) {
-      final textNode = document.getNode(_pastePosition) as TextNode;
-      final pasteTextOffset = (_pastePosition.nodePosition as TextPosition).offset;
-      final attributionsAtPasteOffset = textNode.text.getAllAttributionsAt(pasteTextOffset);
-
-      if (splitContent.length > 1 && pasteTextOffset < textNode.endPosition.offset) {
-        // There is more than 1 node of content being pasted. Therefore,
-        // new nodes will need to be added, which means that the currently
-        // selected text node will be split at the current text offset.
-        // Configure a new node to be added at the end of the pasted content
-        // which contains the trailing text from the currently selected
-        // node.
-        if (currentNodeWithSelection is ParagraphNode) {
-          SplitParagraphCommand(
-            nodeId: currentNodeWithSelection.id,
-            splitPosition: TextPosition(offset: pasteTextOffset),
-            newNodeId: DocumentEditor.createNodeId(),
-            replicateExistingMetadata: false,
-          ).execute(document, transaction);
-        } else {
-          throw Exception('Can\'t handle pasting text within node of type: $currentNodeWithSelection');
-        }
-      }
-
-      // Paste the first piece of content into the selected TextNode.
-      InsertTextCommand(
-        documentPosition: _pastePosition,
-        textToInsert: splitContent.first,
-        attributions: attributionsAtPasteOffset,
-      ).execute(document, transaction);
-
-      // At this point in the paste process, the document selection
-      // position is at the end of the text that was just pasted.
-      newSelectionPosition = DocumentPosition(
-        nodeId: currentNodeWithSelection.id,
-        nodePosition: TextNodePosition(
-          offset: pasteTextOffset + splitContent.first.length,
-        ),
-      );
-
-      // Remove the pasted text from the list of pieces of text
-      // to paste.
-      splitContent.removeAt(0);
-    }
-
-    final newNodes = splitContent
-        .map(
-          // TODO: create nodes based on content inspection.
-          (nodeText) => ParagraphNode(
-            id: DocumentEditor.createNodeId(),
-            text: AttributedText(
-              text: nodeText,
-            ),
-          ),
-        )
-        .toList();
-    _log.log('_PasteEditorCommand', ' - new nodes: $newNodes');
-
-    int newNodeToMergeIndex = 0;
-    DocumentNode mergeAfterNode;
-
-    final nodeWithSelection = document.getNodeById(_pastePosition.nodeId);
-    if (nodeWithSelection == null) {
-      throw Exception(
-          'Failed to complete paste process because the node being pasted into disappeared from the document unexpectedly.');
-    }
-    mergeAfterNode = nodeWithSelection;
-
-    for (int i = newNodeToMergeIndex; i < newNodes.length; ++i) {
-      transaction.insertNodeAfter(
-        existingNode: mergeAfterNode,
-        newNode: newNodes[i],
-      );
-      mergeAfterNode = newNodes[i];
-
-      newSelectionPosition = DocumentPosition(
-        nodeId: mergeAfterNode.id,
-        nodePosition: mergeAfterNode.endPosition,
-      );
-    }
-
-    _composer.selection = DocumentSelection.collapsed(
-      position: newSelectionPosition!,
-    );
-    _log.log('_PasteEditorCommand', ' - new selection: ${_composer.selection}');
-
-    _log.log('_PasteEditorCommand', 'Done with paste command.');
-  }
 }
 
 ExecutionInstruction copyWhenCmdCIsPressed({
@@ -236,13 +70,7 @@ ExecutionInstruction copyWhenCmdCIsPressed({
     return ExecutionInstruction.haltExecution;
   }
 
-  final textToCopy = _textInSelection(
-    document: editContext.editor.document,
-    documentSelection: editContext.composer.selection!,
-  );
-  // TODO: figure out a general approach for asynchronous behaviors that
-  //       need to be carried out in response to user input.
-  _saveToClipboard(textToCopy);
+  editContext.commonOps.copy();
 
   return ExecutionInstruction.haltExecution;
 }
@@ -262,76 +90,9 @@ ExecutionInstruction cutWhenCmdXIsPressed({
     return ExecutionInstruction.haltExecution;
   }
 
-  final textToCut = _textInSelection(
-    document: editContext.editor.document,
-    documentSelection: editContext.composer.selection!,
-  );
-  // TODO: figure out a general approach for asynchronous behaviors that
-  //       need to be carried out in response to user input.
-  _saveToClipboard(textToCut);
+  editContext.commonOps.cut();
 
-  editContext.commonOps.deleteSelection();
   return ExecutionInstruction.haltExecution;
-}
-
-Future<void> _saveToClipboard(String text) {
-  return Clipboard.setData(ClipboardData(text: text));
-}
-
-String _textInSelection({
-  required Document document,
-  required DocumentSelection documentSelection,
-}) {
-  final selectedNodes = document.getNodesInside(
-    documentSelection.base,
-    documentSelection.extent,
-  );
-
-  final buffer = StringBuffer();
-  for (int i = 0; i < selectedNodes.length; ++i) {
-    final selectedNode = selectedNodes[i];
-    dynamic nodeSelection;
-
-    if (i == 0) {
-      // This is the first node and it may be partially selected.
-      final baseSelectionPosition = selectedNode.id == documentSelection.base.nodeId
-          ? documentSelection.base.nodePosition
-          : documentSelection.extent.nodePosition;
-
-      final extentSelectionPosition =
-          selectedNodes.length > 1 ? selectedNode.endPosition : documentSelection.extent.nodePosition;
-
-      nodeSelection = selectedNode.computeSelection(
-        base: baseSelectionPosition,
-        extent: extentSelectionPosition,
-      );
-    } else if (i == selectedNodes.length - 1) {
-      // This is the last node and it may be partially selected.
-      final nodePosition = selectedNode.id == documentSelection.base.nodeId
-          ? documentSelection.base.nodePosition
-          : documentSelection.extent.nodePosition;
-
-      nodeSelection = selectedNode.computeSelection(
-        base: selectedNode.beginningPosition,
-        extent: nodePosition,
-      );
-    } else {
-      // This node is fully selected. Copy the whole thing.
-      nodeSelection = selectedNode.computeSelection(
-        base: selectedNode.beginningPosition,
-        extent: selectedNode.endPosition,
-      );
-    }
-
-    final nodeContent = selectedNode.copyContent(nodeSelection);
-    if (nodeContent != null) {
-      buffer.write(nodeContent);
-      if (i < selectedNodes.length - 1) {
-        buffer.writeln();
-      }
-    }
-  }
-  return buffer.toString();
 }
 
 ExecutionInstruction cmdBToToggleBold({
